requirements.txt
apstra-api-python==0.1.19
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.0.12
idna==2.10
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
requests==2.24.0
urllib3==1.25.11
